48 commits: HQ multi-agent orchestration, desktop UI overhaul,
web API (65+ endpoints), context compaction, 182 E2E tests,
stress-tested CLI with 0% tool error rate across all 9 tools.

+++ crates/ava-tui/src/state/btw.rs
@@ -13,5 +13,5 @@ pub struct BtwState {
     /// Saved messages from before the branch started.
     pub checkpoint_messages: Vec<UiMessage>,
     /// Saved scroll offset from before the branch started.
-    pub checkpoint_scroll: u16,
+    pub checkpoint_scroll: usize,
 }
